Hotfix: refresh expiredIn long으로 변경
Former-commit-id: 57b635ae6efff39db1ff73cfd31bc3f061f29dd9

+++ BE/trip-together/src/main/java/com/ssafy/triptogether/member/data/response/ReissueResponse.java
@@ -11,7 +11,7 @@ public record ReissueResponse(
 	@NotBlank @JsonProperty("access")
 	String access,
 	@NotNull @JsonProperty("expires_in")
-	Integer expiresIn,
+	Long expiresIn,
 	@NotNull @JsonProperty("created_at")
 	Long createdAt
 ) {
